+++ app/src/main/res/layout/activity_layout_one.xml
@@ -11,8 +11,6 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content">
 
-
-
             <com.google.android.material.tabs.TabLayout
                 android:id="@+id/tab_layout"
                 android:layout_width="match_parent"
